requirements.txt
Flask==2.3.3
Flask-SocketIO==5.3.6
python-socketio==5.8.0
python-engineio==4.7.1
openai==1.95.1
python-dotenv==1.0.0
psutil==5.9.6
requests==2.31.0
Werkzeug==2.3.7 